# [Merged by Bors] - improve compile time by type-erasing wgpu structs