--- a/crates/bevy_render/src/render_resource/bind_group_layout.rs
+++ b/crates/bevy_render/src/render_resource/bind_group_layout.rs
@@ -45,6 +45,6 @@ impl Deref for BindGroupLayout {
 
     #[inline]
     fn deref(&self) -> &Self::Target {
-        self.value()
+        &self.value
     }
 }