--- a/crates/bevy_render/src/render_resource/buffer.rs
+++ b/crates/bevy_render/src/render_resource/buffer.rs
@@ -54,7 +54,7 @@ impl Deref for Buffer {
 
     #[inline]
     fn deref(&self) -> &Self::Target {
-        render_resource_ref!(self.value, wgpu::Buffer)
+        render_resource_ref!(&self.value, wgpu::Buffer)
     }
 }