--- a/crates/bevy_render/src/render_resource/mod.rs
+++ b/crates/bevy_render/src/render_resource/mod.rs
@@ -5,6 +5,7 @@ mod buffer_vec;
 mod pipeline;
 mod pipeline_cache;
 mod pipeline_specializer;
+pub mod resource_macros;
 mod shader;
 mod storage_buffer;
 mod texture;