@bjorn3 2022-09-11
I'm curious how this improves compile times. It looks like just as much functions need to be codegened, if not even more for the type erased variants.