@robtfm 2022-09-11
> I'm curious how this improves compile times. It looks like just as much functions need to be codegened, if not even more for the type erased variants.

yes it's weird, and probably something that should be improved in the compiler. as a guess, maybe the size for a struct containing generics needs to be recalculated for each generic combination, even when the generic is not a direct member (like in a box or an arc) and doesn't actually affect the size? the wgpu structs contain some very generic- and associated type-heavy code that is probably causing an exponential type explosion.