@mockersf 2022-09-11
This PR doesn't compile in release on my mac
```
error[E0515]: cannot return reference to local variable `untyped`
   --> crates/bevy_render/src/render_resource/resource_macros.rs:57:9
    |
57  |         &$value
    |         ^^^^^^^ returns a reference to data owned by the current function
    |
   ::: crates/bevy_render/src/render_resource/pipeline_cache.rs:302:9
    |
302 |         render_resource_ref!(untyped, wgpu::PipelineLayout)
    |         --------------------------------------------------- in this macro invocation
    |
    = note: this error originates in the macro `render_resource_ref` (in Nightly builds, run with -Z macro-backtrace for more info)
```