@robtfm 2022-09-11
> This PR doesn't compile in release on my mac

fixed. will address the other feedback tomorrow if there's positive disposition.