@bjorn3 2022-09-12
This seems to be the main contributing factor for bevy_render compiling faster (70% of perf improvement):

```
+-----------------------------------------------------+---------------+------------------+---------------+-------------+------------+------------+--------------+-----------------------+--------------------------+
| Item                                                | Self Time     | Self Time Change | Time          | Time Change | Item count | Cache hits | Blocked time | Incremental load time | Incremental hashing time |
+-----------------------------------------------------+---------------+------------------+---------------+-------------+------------+------------+--------------+-----------------------+--------------------------+
| evaluate_obligation                                 | -5.14754422s  | -90.82%          | -5.147455645s | -88.89%     | -3939      | +0         | +0ns         | +0ns                  | -1.646728ms              |
+-----------------------------------------------------+---------------+------------------+---------------+-------------+------------+------------+--------------+-----------------------+--------------------------+
```

Additionally it saves ~30MB of ~200MB on the file size of of libbevy_render.rlib. Similar results for bevy_pbr (50% of perf improvement):

```
+-----------------------------------------------------+---------------+------------------+---------------+-------------+------------+------------+--------------+-----------------------+--------------------------+
| Item                                                | Self Time     | Self Time Change | Time          | Time Change | Item count | Cache hits | Blocked time | Incremental load time | Incremental hashing time |
+-----------------------------------------------------+---------------+------------------+---------------+-------------+------------+------------+--------------+-----------------------+--------------------------+
| evaluate_obligation                                 | -8.455474125s | -95.79%          | -8.477643881s | -94.68%     | -8198      | +0         | +0ns         | +0ns                  | -2.736299ms              |
+-----------------------------------------------------+---------------+------------------+---------------+-------------+------------+------------+--------------+-----------------------+--------------------------+
```

And saving ~50MB of ~140MB.

(measurements with single threaded compilation and incremental cache cleared)