@MinerSebas 2022-09-22
In addition to the compilation benchmark, a runtime/render benchmark should be performed, so that it is known whether this has any impact here, due to missed optimizations.