@robtfm 2022-09-23
you sent me down a rabbit hole of newtypes not being zero-cost before i read the edit :) ... apparently newtypes are only zero cost as long as they have a non-`calloc` style initialization, which would be true here since we're wrapping an `arc` which needs an initialized counter.