@robtfm 2022-09-29
I thought it was needed so that we can take it in try_unwrap, and not subsequently decrement it on drop.

Maybe there is a better way, I’m not really sure how drop interoperates with a fn(mut self) that destructures the self.

On 28 Sep 2022, at 22:01, Carter Anderson ***@***.******@***.***>> wrote:


@cart commented on this pull request.

________________________________

In crates/bevy_render/src/render_resource/resource_macros.rs<https://github.com/bevyengine/bevy/pull/5950#discussion_r982858371>:

> @@ -0,0 +1,108 @@
+// structs containing wgpu types take a long time to compile. this is particularly bad for generic
+// structs containing wgpu structs. we avoid that in debug builds (and for cargo check and rust analyzer)
+// by boxing and type-erasing with the `render_resource_wrapper` macro.
+// analysis from https://github.com/bevyengine/bevy/pull/5950#issuecomment-1243473071 indicates this is
+// due to `evaluate_obligations`. we should check if this can be removed after a fix lands for
+// https://github.com/rust-lang/rust/issues/99188 (and after other `evaluate_obligations`-related changes).
+#[cfg(debug_assertions)]
+#[macro_export]
+macro_rules! render_resource_wrapper {
+    ($wrapper_type:ident, $wgpu_type:ty) => {
+        #[derive(Clone, Debug)]
+        pub struct $wrapper_type(Option<std::sync::Arc<Box<()>>>);


I'm confused by the Option here. Why is this needed?

—
Reply to this email directly, view it on GitHub<https://github.com/bevyengine/bevy/pull/5950#pullrequestreview-1124352801>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/AMCQEUSU5VDDTOEROX4U3KTWASWZTANCNFSM6AAAAAAQJ5QL3A>.
You are receiving this because you authored the thread.Message ID: ***@***.***>